+++ screen-reader/android/src/main/java/com/capacitorjs/plugins/screenreader/ScreenReader.java
@@ -1,20 +1,8 @@
 package com.capacitorjs.plugins.screenreader;
 
-import com.getcapacitor.JSObject;
-import com.getcapacitor.NativePlugin;
-import com.getcapacitor.Plugin;
-import com.getcapacitor.PluginCall;
-import com.getcapacitor.PluginMethod;
+public class ScreenReader {
 
-@NativePlugin
-public class ScreenReader extends Plugin {
-
-    @PluginMethod
-    public void echo(PluginCall call) {
-        String value = call.getString("value");
-
-        JSObject ret = new JSObject();
-        ret.put("value", value);
-        call.success(ret);
+    public String echo(String value) {
+        return value;
     }
 }
