+++ screen-reader/android/src/main/java/com/capacitorjs/plugins/screenreader/ScreenReaderPlugin.java
@@ -1,0 +1,21 @@
+package com.capacitorjs.plugins.screenreader;
+
+import com.getcapacitor.JSObject;
+import com.getcapacitor.NativePlugin;
+import com.getcapacitor.Plugin;
+import com.getcapacitor.PluginCall;
+import com.getcapacitor.PluginMethod;
+
+@NativePlugin(name = "ScreenReader")
+public class ScreenReaderPlugin extends Plugin {
+    private ScreenReader implementation = new ScreenReader();
+
+    @PluginMethod
+    public void echo(PluginCall call) {
+        String value = call.getString("value");
+
+        JSObject ret = new JSObject();
+        ret.put("value", implementation.echo(value));
+        call.success(ret);
+    }
+}
