+++ screen-reader/ios/Plugin/ScreenReader.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+@objc public class ScreenReader: NSObject {
+// swiftlint:disable:previous type_name
+    @objc public func echo(_ value: String) -> String {
+        return value
+    }
+}
