+++ screen-reader/ios/Plugin/ScreenReaderPlugin.m
@@ -3,6 +3,6 @@
 
 // Define the plugin using the CAP_PLUGIN Macro, and
 // each method the plugin supports using the CAP_PLUGIN_METHOD macro.
-CAP_PLUGIN(ScreenReader, "ScreenReader",
+CAP_PLUGIN(ScreenReaderPlugin, "ScreenReader",
            CAP_PLUGIN_METHOD(echo, CAPPluginReturnPromise);
 )
