+++ screen-reader/ios/Plugin/ScreenReaderPlugin.swift
@@ -5,13 +5,15 @@ import Capacitor
  * Please read the Capacitor iOS Plugin Development Guide
  * here: https://capacitorjs.com/docs/plugins/ios
  */
-@objc(ScreenReader)
-public class ScreenReader: CAPPlugin {
+@objc(ScreenReaderPlugin)
+public class ScreenReaderPlugin: CAPPlugin {
+// swiftlint:disable:previous type_name
+    private let implementation = ScreenReader()
 
     @objc func echo(_ call: CAPPluginCall) {
         let value = call.getString("value") ?? ""
         call.success([
-            "value": value
+            "value": implementation.echo(value)
         ])
     }
 }
