+++ screen-reader/package.json
@@ -31,7 +31,7 @@
     "typescript": "~3.9.7"
   },
   "peerDependencies": {
-    "@capacitor/core": "^2.3.0"
+    "@capacitor/core": "^3.0.0-alpha.0"
   },
   "files": [
     "dist/",
